+++ spec/integration/knife/chef_fs_data_store_spec.rb
@@ -22,7 +22,7 @@ require 'chef/knife/show'
 require 'chef/knife/raw'
 require 'chef/knife/cookbook_upload'
 
-describe 'knife raw -z' do
+describe 'ChefFSDataStore tests' do
   extend IntegrationSupport
   include KnifeSupport